Set up ACLK correctly
This fixes two issues: rand() is no longer very slow and the discrepancy
between running the test with mspdebug and without it is gone. Prior to
this change, the tests would be slow when run in mspdebug and would hang
indefinitely when run outside it. Now they take a reasonable amount of
time in both cases.